subscription and invite friend

+++ Apps/paisa2u/paisa2u.common/Models/Subscriptionperc.cs
@@ -9,13 +9,13 @@ public partial class Subscriptionperc
 
     public int RegId { get; set; }
 
-    public float? Appowner { get; set; }
+    public double? Appowner { get; set; }
 
-    public float? Vendor { get; set; }
+    public double? Vendor { get; set; }
 
-    public float? Subvendor { get; set; }
+    public double? Subvendor { get; set; }
 
-    public float? Customer { get; set; }
+    public double? Customer { get; set; }
 
     public DateTime Endate { get; set; }
